Refactoring aliasing.
