+++ package.json
@@ -11,7 +11,7 @@
     "eslint-check": "eslint --print-config src/components/app/container/index.js | eslint-config-prettier-check",
     "lint": "npm run lint:css && npm run lint:js",
     "lint:css": "stylelint src/**/*.scss",
-    "lint:js": "eslint --ext js --ext mjs server/ src/ test-support/ webpack/",
+    "lint:js": "eslint --ext js --ext mjs server/ src/ support/ webpack/ jest.config.js postcss.config.js",
     "pretest": "npm run lint",
     "start": "node --experimental-modules server/index.mjs",
     "test": "jest",
