+++ src/components/loader/tests/index.spec.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import Loader from '../';
 
-jest.mock('../../outsiders-mark', () => 'MockOutsidersMark');
+jest.mock('Components/outsiders-mark', () => 'MockOutsidersMark');
 
 describe('Loader component', () => {
   let testContext;
