+++ support/tests/file-transformer.js
