--- webpack/paths.js
@@ -1,9 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  root: path.join(__dirname, '../../'),
-  dist: path.join(__dirname, '../dist'),
-  images: path.join(__dirname, '../images/'),
-  server: path.join(__dirname, '../server/'),
-  powerImages: path.join(__dirname, '../images/powers')
-};
